packages/ui restructuring
