+++ apps/web/app/page.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import { HomeScreen, Provider } from '@u3/ui';
+import { PrimitivesSample, UIProvider } from '@u3/ui';
 import type React from 'react';
 
 export default function HomePage(): React.ReactElement {
   return (
-    <Provider>
+    <UIProvider>
       <div style={{ minHeight: '100vh', flex: 1 }}>
-        <HomeScreen />
+        <PrimitivesSample />
       </div>
-    </Provider>
+    </UIProvider>
   );
 }
