--- packages/ui/src/components/index.ts
@@ -1,8 +1,0 @@
-export type { ButtonProps } from './Button';
-export { Button } from './Button';
-export type { CardProps } from './Card';
-export { Card } from './Card';
-export type { InputProps } from './Input';
-export { Input } from './Input';
-export type { TextProps } from './Text';
-export { Text } from './Text';
