+++ packages/ui/src/layout/PrimitivesSample.tsx
@@ -1,10 +1,10 @@
 import { StatusBar } from 'expo-status-bar';
 import { Paragraph, ScrollView, SizableText, View, YStack } from 'tamagui';
-import { Button } from '../../components/Button';
-import { Card } from '../../components/Card';
-import { Input } from '../../components/Input';
+import { Button } from '../primitives/Button';
+import { Card } from '../primitives/Card';
+import { Input } from '../primitives/Input';
 
-export function HomeScreen() {
+export function PrimitivesSample() {
   return (
     <>
       <StatusBar style='light' />
